Fix release scripts

+++ scripts/release_prepare.sh
@@ -1,18 +1,21 @@
 #!/bin/bash
 
-SCRIPTDIR="$( cd -- "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
-BASEDIR="$( cd -- "$(dirname "$SCRIPTDIR")" >/dev/null 2>&1 ; pwd -P )"
+SCRIPT_DIR="$( cd -- "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
+PLUGIN_DIR="$( cd -- "$(dirname "$SCRIPT_DIR")" >/dev/null 2>&1 ; pwd -P )"
+PLUGINS_ROOT_DIR="$( cd -- "$(dirname "$PLUGIN_DIR")" >/dev/null 2>&1 ; pwd -P )"
+PLUGIN_SLUG=$(basename $PLUGIN_DIR)
 
-cd $BASEDIR
-rm -rf vendor/
-composer install --no-dev
-cd ..
+if [[ -f "$PLUGIN_DIR/composer.json" ]]; then
+  rm -rf "$PLUGIN_DIR/vendor"
+  composer install --no-dev
+fi
 
-rm digital-license-manager.zip
+rm "$PLUGINS_ROOT_DIR/$PLUGIN_SLUG.zip"
+cd $PLUGINS_ROOT_DIR
 
-zip -r wp-mkd-air-quality.zip wp-mkd-air-quality \
--x "wp-mkd-air-quality/.git/*" \
--x "wp-mkd-air-quality/.gitignore" \
--x "wp-mkd-air-quality/README.md";
+zip -r "$PLUGIN_SLUG.zip" "$PLUGIN_SLUG" \
+ -x="*scripts*" \
+ -x="*.git*" \
+ -x="README.md"
 
 echo "New version ready."
